requirements.txt
Flask==3.1.0
pyenchant==3.2.2
pyflowchart==0.3.1
waitress==3.0.2
python-dotenv==1.0.1